Support displaced grass around characters.
This is a huge commit because I didn't really stop until I had solved the problem.

The main thrust of this work is adding support for "Displacement Masks". These are custom render buffers that coexist with the main screen buffer and certain elements in the game render to this mask exclusively. The first and principle use-case for a displacement mask is to capture where on the screen there are entities which should push/bend/depress grass around them when overlapping.

I've tried to document how these things work to the best of my ability (and understanding). There's no single reference I can cite - this is sort of a standard practice in computer graphics (rendering to a virtual texture instead of directly to the screen) that we're adapting for this purpose.

Non-exhaustive accounting of changes:
- Added /Docs to the .csproj so we can start including Markdown files for things that get too confusing to remember.
- Added /Docs/Rendering.md and attempted to document some high-level notes about both Displacement Masks and Grass rendering.
- Added a new SubViewport type "DisplacementMaskViewport" and an associated scene file. These know how to create and maintain the Displacement Mask virtual texture at runtime.
- Added an instance of "DisplacementMaskViewport" as a new global viewport on the Main scene.
- Added 2 new global shader uniforms.
-- `screen_world_rect` can be used to access the global position and size of the screen in world coordinates from a shader.
-- `displacement_mask_tex` can be used to access the virtual texture rendered by the DisplacementMaskViewport each frame.
- All Characters now instance and register a "GrassDisplacementMarker" with the global DisplacementMaskViewport while alive.
- Wrote/rewrote the entire grass_patch.gdshader to sample from the displacement mask, bend grass away from characters, sway in the wind, have transparency, etc.
-- (There is SO much more to do here...)
- Changed the clear / background color of our project from default gray to pure black. This was sort of necessary in order for the DisplacementMaskViewport's rendering to avoid having to write all black to the texture each frame before rendering sprites. We can revisit this if it somehow becomes an issue.