Move grass displacement logic to Moveable.cs
All Moveables now support grass displacement.

Currently there is no logic for changing the size of the displacement so
all projectiles displace the same way that characters do. This will take
a bit of work to get right because there is nuance to projectile
displacement (it technically should not displace based on the base of
the grass, as projectiles are in the air).