Decouple DisplacementMask color from global clear color.
DisplacementMaskViewport now creates a Sprite2D under the `MarkerRoot`
container node with the name `ClearColorSprite`. This sprite is
dynamically-resized to fit the viewport each frame and generates a 1x1
pixel texture with a constant color controlled by `ClearColor` on the
Viewport node.

With the above, the project's default clear color has been reverted to
the editor's preference.

This also includes some minor work on DisplacementMaskMarker for
controlling what color is set on the mask, but it's currently unused.