Prevent drag-resizing multiple RectRegions at once.
This was possible before if the control points overlapped while dragging
one to resize a RectRegion that overlapped another.

This also may fix a periodic editor crash that is clearly related to the
above, but I haven't ever gotten a stack trace out when it happens so
I'm not positive if I've fixed it here. Seems likely, though.