Change displacement markers to additive blending.
Added a CanvasItemShader to each marker sprite that switches it from
Mixed blending to Additive blending. Since the sprites are gradient
circles with a black background, the black edges were overwriting the
white circles of other markers on the mask (Mix). Switching to additive
blending means that white (1) and black (0) make white (1).

This is more important for the WIP displacement trails work I'm doing.
But it does have a minor impact on rendering quality when lots of
markers are near each other in the grass.