Add basic trails through the grass.
This is heavily WIP but I wanted to check something in that gets the
idea in place with minimal added work so that it's easy to remove if we
don't like it.

This adds a new type of viewport, a FrameBufferViewport, which creates a
full-viewport sprite that renders the same thing that ANOTHER viewport
is already rendering. This is the only way that Godot supports multiple
render targets right now.

So we now have 3 viewports dedicated to displacement. We don't NEED 3
but we have 3 right now:
1. Screen-space Displacement Viewport: it only works for the current
   screen. If GrassPatch.cs is configured with
   UseGlobalDisplacementMask=false, the Screen-space viewport is where
   displacement data comes from.
2. Global Displacement Viewport: this covers the entire world with a
   (currently 512x512) displacement mask. This is what gets used when
   GrassPatch.UseGlobalDisplacementMask=true.
3. Global Displacement FrameBuffer Viewport: this mirrors the global
   displacement viewport each frame.

How trails work:
- The Global Displacement FrameBuffer Viewport tracks the Global
  Displacement Viewport. It renders the same thing that viewport sees.
- The Global Displacement Viewport is configured with a FrameBuffer
  reference to the Global Displacement FrameBuffer viewport. This
  FrameBuffer reference gets assigned to a fullscreen sprite which
  renders on top of the clear color, but below all displacement markers.
- - This causes the Global Displacement Viewport to start with the last
  frame's result as a baseline, because of the sprite rendering the
  FrameBuffer viewport's contents.

And that's actually it. Trails will naturally accumulate on the Global
Displacement Viewport each frame as the previous frame is layered on
first and new marker data is added on top.

Work still to be done:
- The Global Displacement Viewport should use a shader to sample the
  previous frame. This would let us do things like discard certain
  colors of displacement (e.g. projectiles) and only keep the trail data
  for characters. We can also use this to automatically fade away old
  displacement data.
- You can't use a FrameBuffer without trails right now. There could be a
  usecase for that. (This is probably solved by the shader work just
  mentioned above).