Grass stands back up on its own after time.
Added the ability for DisplacementMaskViewports to configure a
RefreshShader that is responsible for handling updates to the mask
BEFORE any markers are rendered that frame.

In its most basic form this shader would do the same thing that the
current code does: render the last frame's displacement mask (configured
on the FrameBuffer) to a sprite on the GlobalDisplacementViewport each
frame. This creates the "trails".

The shader allows us to do more advanced effects. The current shader
(update_displacement_mask.gdshader) configured on
GlobalDisplacementViewport implements a gradual fade-out effect for any
data on the displacement mask. The fade-out causes trails to gradually
disappear over time, which in turn causes the grass to stand back up as
the displacement approaches 0.