Fix age-weighted grass displacement math.
This was all sorts of wrong and I just didn't really notice. The age
wasn't applying to the actual magnitude of displacement, it was getting
wrongly applied to the strength of the gradient, which barely affects
anything at all if you read through the shader.

Fixed it to apply to the displacement magnitude and now it looks/behaves
right for different shades of Blue displacement.