Make projectiles displace grass ~75% less.
This changes the color of the default projectile displacement marker to
have a slight yellow tint due to its Blue values being consistently
lower (the max right now is ~0.27). Because the grass shader interprets
Blue as the relative age of the displacement, this causes projectiles to
1) displace grass for less time, and 2) displace grass to a lesser
   degree.

The current Blue value was chosen just by trying a few and observing the
grass behavior.