Stop packing blade midpoint into GREEN vertex color.
This doesn't always work well due to the displacement masks using HDR
textures. There's a way to make color-packing work there (and it's
actually still safe to use for the Phase in RED) but I need to read more
about texture formats and sampling procedures.

In the meantime, changed the grass mesh generation to just put all
vertices for the same blade of grass at the same X/Y location. The
location chosen is the midpoint of each blade on X, at its lowest Y
position.

The grass_patch shader has been updated to extrude/grow the grass blade
in the vertex shader, based on two new uniforms `blade_width` and
`blade_height`.