Rewrite grass displacement to support directional data.
This makes minor improvements to the way grass gets displaced, but
really it's a bunch of groundwork for us to support velocities in the
displacement map.

DisplacementMaskViewport now accepts an override for the shader used to
render any child Displacement Markers. Displacement Markers themselves
are opaque or gradient shapes that express *intensity* of displacement
at some pixel in the world.

The new shaders that run over the displacement markers translate the
marker data into whatever information that particular
DisplacementMaskViewport needs. Right now, we only use these for grass,
so the shader associated with our global displacement viewport
(update_displacement_marker.gdshader) handles it for grass.

The new marker shader multiplies every intensity point on the marker by
the directional offset from the center, and encodes the result in the
red/green channels of the marker. So for a circular marker like we have
currently, points farther out will be lower intensity (sourced from the
marker sprite) and then will be multiplied in the direction that they
are from the center of the marker sprite (e.g. left of center => (-1,
0)) to produce a red/green pixel color which captures the direction AND
  intensity.

A lot of math had to be adjusted for this change, mostly caused by a
wild goose chase while I tracked down an erroneous clamp() call within
the fade-out logic of update_displacement_mask.gdshader.

Many textures and resources related to the displacement masks were also
updated to use Nearest filtering, to avoid any unexpected blending
during shader samples.