Constrain grass bending to maintain blade length.
The grass was "bending" to positions that were longer than the blade was
supposed to be as determined by blade_height. Added some (admittedly
imperfect) math to clamp the range that the grass can displace to so
that it doesn't stretch *too* much (some stretch is still allowed).

This change also tweaks the size and resolution of the character
displacement marker so that it bends grass a bit more finely. The
resolution of the Global Displacement Mask (and its FrameBuffer) have
been increased to 1024x1024 which has a major impact on the quality of
grass displacement as well.