Add grass painter node.
This is experimental right now. GrassPainter is what is sometimes called
a "toroidal array" meaning it wraps around on all edges. The array is 2D
and stores rows of grass that cover the whole screen. The amount of
grass owned by this node is slightly more than what is necessary to
cover the screen so as the viewport moves around the world, the grass
stays visibly stationary while off-screen it is being constantly
repositioned to fill newly-exposed areas.

The end result is a scrolling tile of "infinite" grass rows that cover
the entire world.

Right now this is super crude and the grass *literally* covers the
*entire* world (including walls and doors) so the painter node is marked
invisible. But it's there... scrolling... lurking...