Increase grass blade size, reduce overall count.
Also added shader parameter for controlling how pointy the grass is at
each blade tip. The default (1.0) is a perfect point i.e. the blade is a
triangle. At 0, the blade is a rectangle.

Tweaked the number of blades and the size of each blade on the test
grasspatch. Also tweaked the size of the blades and size of each patch
for the grass painter (still invisible).