Prevent crashing the editor when dragging objects over RectRegions.
RectRegion abuses the single-process nature of the editor and game and
tries to do fancy things with input processing. The editor sees input
first, so we can't use traditional inputs like left/right/middle mouse
for click-and-drag.

Addeed an additional protection check for the whole thing and also moved
click-and-drag to MouseButton X1. This is normally the first side-button
on your mouse. Not sure what to do for trackpads...