Make AI Brains a full Node2D instead of a Resource.
I had made these resources so that they could be defined inline as a
property on the NPC. Now that I'm starting to think about adding new
types of NPC it seems unnecessary to have to implement a subclass of
NonPlayerCharacter just for an NPC to think differently than another NPC
with otherwise the same movement logic and whatnot.

So now Brains can be subclassed as nodes that are associated with each
NonPlayerCharacter. No functional change yet though.