Fix editor event subscription mismatch.
When you run [Tool] code in the editor, it calls the internal node
functions differently:

_Ready() only gets called once when the node is loaded and added to a
scene for the first time.

_EnterTree() is called any time the node's scene tree becomes relevant
in the editor again (such as when you select the tab for that scene).

_ExitTree() is called any time the node's scene tree stops being
relevant (you tab away in the editor for example).

We were subscribing to Editor SelectionChanged events during _Ready(),
and then unsubscribing during _ExitTree(). So the event was subscribed
to once the very first time the RadialSpawner node was added to a scene,
and then it was removed N times (where N = number of times the scene's
tab was viewed in the editor and then swapped away).

Changed the subscription to be symmetric during _EnterTree()/_ExitTree()
so that the same number of both actions occur. With logging enabled this
can be seen in the editor as you swap tabs, but that logging is
commented out.