Add level transition mechanism
Placed stairs in the middle building that let the player transition
between levels

The old level is paused and hidden in the scene tree, so when the player
returns everything is exactly as it was left. Later this should
probably be replaced with a save/load system

Temporarily disabled camera smoothing, since this makes the camera jumpy
during level transitions. I could just disable/renable it during the
transition (which works well), but I want to rethink camera smoothing,
as I think it contributes to the sprite jitter/tearing artifacts