Automatically add outdoor static bodies to FLOOR 0.
This is a short-term fix where elements in the scene that aren't inside
a building get automatically added to the first floor's nav map baking
group (FLOOR 0).

Note: Right now SmallTown.cs places procedural geometry AFTER navmap
generation. This is intentional. Baking the navmap for the forest is
expensive and takes ~30 seconds. Need to come up with a better approach.