Refactor NPC RotationGoal as generic Character.LookAtAngle
Remove NonPlayerCharacter's RotationGoal and replace it with LookAtAngle on Character instead. These conceptually serve the same purpose but the latter better captures what it's for and how it's used.

Previously, RotationGoal (now LookAtAngle) was exclusively used to control how fast NPCs would rotate when turning. This worked okay for top-down NPCs like Shambler who could use the same Sprite rotated any angle.

For Characters like Player or other NPCs with animated sprites, this doesn't make as much sense. It very likely won't be the case that we want the LookAtAngle applied to the GlobalRotation (Character.RotateTowardsLookAt=false disables this), and we can't / don't want to try and derive the sprite animations from GlobalRotation anyway. For animated characters we want to use either their velocity or some other indicator of what direction the sprite should be facing - and that is where LookAtAngle comes in.

LookAtAngle has other potential uses:
1. If we introduce characters who can move their head to look at things without moving their body.
2. If we want the existing behavior of delaying GlobalRotation's movement to a new angle, such as for a 360 degree turret or laser.
3. Deriving visibility cones that move with the gaze of an NPC who is otherwise stationary.