(WorldLifecycleManager) Only call PostWorldInit once.
I left some debug logging in the AI Brain code accidentally and noticed
it was logging more than it should be. Tracked the bug back to WLM
calling PostWorldInit() multiple times on nodes that are re-parented in
the tree because they technically _EnterTree() multiple times,
triggering the manager's hooks to fire.

Added some idempotency logic to WLM as a stop-gap. Could do a lot more
here but this should be sufficient for now.