Support move/idle animations for characters.
Adding basic support for characters with AnimatedSprite2D sprites to
automatically play movement and idle animations based on naming
convention.

Movement animations are expected to be in the form
`<type>_<YDirection>_<XDirection>` where the directional components are
optional, e.g.:
- `idle_up` (no XDirection)
- `idle_right` (no YDirection)
- `idle_up_right` (both present)

Two types of animation are supported: `idle` and `move`. The choice to
play one set of animations over the other comes down to whether the
character currently has a non-zero velocity.

Optionally, characters may be configured to derive their `idle`
animation directions from the LookAtAngle of the character. Player uses
this, as the LookAtAngle represents where the player is aiming. The
default is `true` because most characters in our games will have 6-8DOF
animations and won't use GlobalRotation most likely anyway.

The Player.tscn has been updated with an AnimatedSprite2D that has all
of these idle/move animations defined so that it can be demonstrated in
the test worlds.